Remove the examples on nonce derivation
This is just an XOR operation.

The intent was to highlight the fact that nonce sizes in AEGIS
differ from other AEADs, and that the code responsible for nonce
derivation was likely affected in implementations.

But that can be mentioned in the rest of the document.

Previous examples are enough to show that AEGIS uses 128-bit and
256-bit nonces.